pyproject.toml
[project]
name = "agentic-graphrag-med"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.3.5",
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.0",
    "langchain-groq>=1.1.0",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=1.0.4",
    "neo4j-driver>=5.28.2",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pypdf>=6.4.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
]
